separate out SettingViewModel and SettingsViewModel
